requirements.txt
streamlit==0.76.0
opencv_python==4.5.1.48
Keras==2.4.3
numpy==1.20.1
Pillow==8.1.2
